WIP refactoring decentraland util functions
